+++ apps/web/app/(auth)/sign-in/page.tsx
@@ -17,7 +17,7 @@ import { useForm } from 'react-hook-form';
 import { PasswordInput } from '@/app/(auth)/_components/password-input';
 import { SocialAuthButtons } from '@/app/(auth)/_components/social-auth-buttons';
 import { useAuth } from '@/hooks/use-auth';
-import { type SignInValues, signInSchema } from '@/validations/auth';
+import { type SignInValues, signInSchema } from '@/validations/auth.schema';
 
 export default function SignInPage() {
   const router = useRouter();
